+++ src/tests/wallet_test.ts
@@ -1441,8 +1441,12 @@ describe("Wallet Class", () => {
       jest.spyOn(Wallet.prototype, "createWebhook").mockReturnValue(wh);
       const result = await wallet.createWebhook("https://example.com/callback");
       expect(result).toBeInstanceOf(Webhook);
-      expect((result.getEventTypeFilter() as WebhookWalletActivityFilter)?.wallet_id).toBe(walletModel.id);
-      expect((result.getEventTypeFilter() as WebhookWalletActivityFilter)?.addresses).toStrictEqual([address1]);
+      expect((result.getEventTypeFilter() as WebhookWalletActivityFilter)?.wallet_id).toBe(
+        walletModel.id,
+      );
+      expect((result.getEventTypeFilter() as WebhookWalletActivityFilter)?.addresses).toStrictEqual(
+        [address1],
+      );
       expect(result.getEventType()).toBe("wallet_activity");
     });
   });